spostato tutti i metodi
