+++ src/com/iaco2code/springdemo/daoevent/EventDAO.java
@@ -1,14 +1,21 @@
 package com.iaco2code.springdemo.daoevent;
 
+import java.util.List;
+
 import com.iaco2code.springdemo.entity.Evento;
 
 public interface EventDAO {
 
-	void saveEvent(Evento theEvent);
+	public void saveEvent(Evento theEvent);
 
-	void deleteEvent(Evento theEvent);
+	public void deleteEvent(Evento theEvent);
 
-	Evento getEventId(int idEvent);
+	public Evento getEventId(int idEvent);
 	
-	String getTipoEvent(int idEvent);
+	public String getTipoEvent(int idEvent);
+	
+	public List<Evento> getEvent(String evento);
+	
+	public List<Evento> getEvents();
+
 }
