+++ src/com/iaco2code/springdemo/daoevent/EventDAOImpl.java
@@ -79,4 +79,40 @@ public class EventDAOImpl implements EventDAO {
 				return null;	
 			}
 	}
+	
+	@Override
+	@Transactional
+	public List<Evento> getEvent(String evento) {
+		
+		//get the current hibernate session
+		Session currentSession = sessionFactory.getCurrentSession();
+		
+		//create query
+		Query<Evento> theQuery = currentSession.createQuery("from Evento where Tipo='"+evento+"'");
+		
+		
+		//execute query and get result list
+		List<Evento> eventi = theQuery.getResultList();
+		
+		//return the result
+		return eventi;
+		
+	}
+	
+	@Override
+	@Transactional
+	public List<Evento> getEvents() {
+		//get the current hibernate session
+		Session currentSession = sessionFactory.getCurrentSession();
+			
+		//create query
+		Query<Evento> theQuery = currentSession.createQuery("from Evento");
+						
+				
+		//execute query and get result list
+		List<Evento> eventi = theQuery.getResultList();
+				
+		//return the result
+		return eventi;
+	}
 }
